Created artist/id route and Top songs
had to add client credential flow to allow us to just use the public part of spotifys api, the auth part is still in there and should be used when pertaining to a certain user.

Top-50 songs are pulled from spotifys global play list.

for now set a client id and client secret on your own pc when we deploy we will have to save it to the server somewhere not plain text visible